+++ core/events/party.go
@@ -15,20 +15,20 @@ package events
 import (
 	"context"
 
+	"code.vegaprotocol.io/vega/core/types"
 	proto "code.vegaprotocol.io/vega/protos/vega"
 	eventspb "code.vegaprotocol.io/vega/protos/vega/events/v1"
 )
 
 type Party struct {
 	*Base
-	p proto.Party
+	p types.Party
 }
 
-func NewPartyEvent(ctx context.Context, p proto.Party) *Party {
-	cpy := p.DeepClone()
+func NewPartyEvent(ctx context.Context, party types.Party) *Party {
 	return &Party{
 		Base: newBase(ctx, PartyEvent),
-		p:    *cpy,
+		p:    party,
 	}
 }
 
@@ -36,18 +36,18 @@ func (p Party) IsParty(id string) bool {
 	return p.p.Id == id
 }
 
-func (p *Party) Party() proto.Party {
+func (p *Party) Party() types.Party {
 	return p.p
 }
 
-func (p Party) Proto() proto.Party {
-	return p.p
+func (p Party) Proto() *proto.Party {
+	return p.p.IntoProto()
 }
 
 func (p Party) StreamMessage() *eventspb.BusEvent {
 	busEvent := newBusEventFromBase(p.Base)
 	busEvent.Event = &eventspb.BusEvent_Party{
-		Party: &p.p,
+		Party: p.Proto(),
 	}
 
 	return busEvent
@@ -56,6 +56,6 @@ func (p Party) StreamMessage() *eventspb.BusEvent {
 func PartyEventFromStream(ctx context.Context, be *eventspb.BusEvent) *Party {
 	return &Party{
 		Base: newBaseFromBusEvent(ctx, PartyEvent, be),
-		p:    *be.GetParty(),
+		p:    types.Party{Id: be.GetParty().Id},
 	}
 }